add `relative_url` filter for image link
Since our site uses a baseurl (i.e. the `infolit-for-everyone` in the address bar), we need to tell GitHub Pages not to just look in the main `idaho-infolit` organization when we start a link with `/`.

Here's the pattern:
![my image]({{ "/assets/images/my-image.png" | relative_url }})

You can find more info about this in [Minimal Mistakes Issue #2510](https://github.com/mmistakes/minimal-mistakes/issues/2510).